+++ src/py42/clients/storage/archive.py
@@ -18,17 +18,19 @@ class StorageArchiveClient(BaseClient):
         Args:
             session_id (int): The ID for the web restore session.
             device_guid (str): The GUID for the device.
-            regex (str, optional): Filename regex to filter results by. Defaults to None.
+            regex (str, optional): A filename regex to filter results by. Defaults to None.
             max_results (int, optional): The max results to return. Defaults to None.
-            timestamp (int, optional): The POSIX timestamp of the archive against which to
-                search. 0 indicates the most recent version.
-            show_deleted:
+            timestamp (int, optional): The POSIX timestamp (seconds) of the archive against which
+                to search. 0 indicates the most recent version. It will return all versions older
+                than the timestamp you provide. Defaults to None.
+            show_deleted (bool, optional): Set to True to include deleted files in the search.
+                Defaults to None.
 
         Returns:
-
+            :class:`py42.sdk.response.Py42Response`
         """
-        # session_id is a web restore session ID (see create_restore_session)
         uri = u"/api/WebRestoreSearch"
+        timestamp = timestamp * 1000 if timestamp else timestamp
         params = {
             u"webRestoreSessionId": session_id,
             u"guid": device_guid,