+++ CHANGELOG.md
@@ -16,6 +16,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
     instead of ("ACTIVE", "INACTIVE", or "ALL") on the following `LegalHoldClient` methods:
     - `get_all_matters()`
     - `get_all_matter_custodians()`
+- Parameter `timestamp` for `StorageArchiveClient.search_paths()` now uses seconds instead of milliseconds.
 
 ### Removed